backend/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
fpdf2
google-generativeai
boto3
botocore
regex
pytest
httpx
